Create demux.v5.bash
Script that initiate demultiplexing through the Casava pipeline.
After demultiplexing it will parse some quality values from Casava to a database at the database server using the python script parseunaligned_dbserver.py. This info is then retrieved using selectunaligned_dbserver.py to a statistics log that is stored together with demultiplexed data.

This script is also generating a log that is sent with the data. It should be used to record versions of script together fail/pass values . . .